Make review consistent even if new code is pushed
#### Summary
Before this change code review was applied to the latest commit assuming
that no code had change since it was fetched. After this change the code
review is applied to the correct commit (the one that was pulled), and
even if the PR changes during review the comments end up in the right
place.

#### Test Plan
- [ ] Manual test